refactor(core): Move decorator and metadata service to @n8n/decorators
Address PR review feedback from afitzek:

- Move @BreakingChangeRule decorator and BreakingChangeRuleMetadata service
  to @n8n/decorators package, following the established @BackendModule /
  ModuleMetadata pattern. This allows rules to be defined anywhere in the
  codebase, not just within the breaking-changes module.
- Fix discovery test to assert exactly 17 rules instead of just > 0,
  catching the case where getEntries() returns an empty array.
