+++ packages/@n8n/decorators/src/breaking-change-rule/index.ts
@@ -1,0 +1,2 @@
+export { BreakingChangeRule } from './breaking-change-rule';
+export { BreakingChangeRuleMetadata } from './breaking-change-rule-metadata';
