+++ packages/@n8n/decorators/src/index.ts
@@ -1,3 +1,4 @@
+export * from './breaking-change-rule';
 export * from './controller';
 export * from './command';
 export { Debounce } from './debounce';
