+++ packages/cli/src/modules/breaking-changes/rules/v2/task-runner-docker-image.rule.ts
@@ -1,6 +1,6 @@
 import { GlobalConfig } from '@n8n/config';
 
-import { BreakingChangeRule } from '../../breaking-change-rule.decorator';
+import { BreakingChangeRule } from '@n8n/decorators';
 import type {
 	BreakingChangeRuleMetadata,
 	IBreakingChangeInstanceRule,
