+++ packages/cli/src/modules/breaking-changes/rules/v2/tunnel-option.rule.ts
@@ -1,4 +1,4 @@
-import { BreakingChangeRule } from '../../breaking-change-rule.decorator';
+import { BreakingChangeRule } from '@n8n/decorators';
 import type {
 	BreakingChangeRuleMetadata,
 	IBreakingChangeInstanceRule,
